+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Next
 
+* Switch from `action` to `navigation` (which contains an `action` as well as the previous `response`).
+  * Make `navigation` a required prop of the `<CuriBase>` (`action` was optional).
+  * `<Curious>` injects `navigation`, not `action`.
 * `<Curious>` sets initial `response`/`action` when responsive.
 
 ## 1.0.0-alpha.2